# openapi.yaml
openapi: 3.0.3
info:
  title: Evenly API
  version: 0.1.0
  description: |
    HTTP contract expected by the frontend service adapter. Money values are
    integer minor units (kobo for NGN). The frontend sends and receives JSON
    except for receipt uploads, which use multipart/form-data.
servers:
  - url: /api
    description: Configure the frontend base URL with VITE_API_BASE_URL.
...
paths:
...
  /groups:
    get:
      tags: [Groups]
      summary: List accessible saved groups
      security: []
      description: Authentication is required for saved groups; a guest receives an empty list.
      responses:
        '200': { description: Group summaries, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/GroupSummary' } } } } }
    post:
      tags: [Groups]
      summary: Create a group
      security: [{ sessionCookie: [] }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/CreateGroupInput' } } } }
      responses:
        '201': { description: Created group, content: { application/json: { schema: { $ref: '#/components/schemas/Group' } } } }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /groups/{groupId}:
    parameters: [{ $ref: '#/components/parameters/GroupId' }]
    get:
      tags: [Groups]
      summary: Get group details
      security: []
      description: Requires membership for saved groups, or guest ownership/access to the quick split. Authentication is required for saved groups.
      responses:
        '200': { description: Group detail, content: { application/json: { schema: { $ref: '#/components/schemas/GroupDetail' } } } }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [Groups]
      summary: Update group details
      security: [{ sessionCookie: [] }]
      description: Requires group-admin access (quick-split owner is also an admin).
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/UpdateGroupInput' } } } }
      responses:
        '200': { description: Updated group, content: { application/json: { schema: { $ref: '#/components/schemas/Group' } } } }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /quick-splits:
    post:
      tags: [Groups]
      summary: Create a quick split
      security: []
      description: Public; the first member represents the viewer.
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/CreateQuickSplitInput' } } } }
      responses:
        '201': { description: Created quick split, content: { application/json: { schema: { $ref: '#/components/schemas/Group' } } } }
        '422': { $ref: '#/components/responses/ValidationError' }
  /groups/{groupId}/save:
    parameters: [{ $ref: '#/components/parameters/GroupId' }]
    post:
      tags: [Groups]
      summary: Save a quick split as a group
      security: [{ sessionCookie: [] }]
      responses:
        '200': { description: Saved group, content: { application/json: { schema: { $ref: '#/components/schemas/Group' } } } }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
  /groups/{groupId}/members:
    parameters: [{ $ref: '#/components/parameters/GroupId' }]
    post:
      tags: [Groups]
      summary: Add a member by display name
      security: [{ sessionCookie: [] }]
      description: Requires group-admin access.
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/AddMemberInput' } } } }
      responses:
        '201': { description: Added member, content: { application/json: { schema: { $ref: '#/components/schemas/GroupMember' } } } }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /groups/{groupId}/members/{memberId}:
    parameters:
      - { $ref: '#/components/parameters/GroupId' }
      - { $ref: '#/components/parameters/MemberId' }
    delete:
      tags: [Groups]
      summary: Remove a member
      security: [{ sessionCookie: [] }]
      description: Requires group-admin access.
      responses:
        '204': { description: Member removed }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /groups/{groupId}/invite:
    parameters: [{ $ref: '#/components/parameters/GroupId' }]
    post:
      tags: [Invites]
      summary: Create or rotate an invite link
      security: [{ sessionCookie: [] }]
      description: Requires group-admin access.
      responses:
        '200': { description: Invite token, content: { application/json: { schema: { $ref: '#/components/schemas/InviteToken' } } } }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
    delete:
      tags: [Invites]
      summary: Revoke an invite link
      security: [{ sessionCookie: [] }]
      description: Requires group-admin access.
      responses:
        '204': { description: Invite revoked }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /invites/{token}:
    parameters: [{ $ref: '#/components/parameters/InviteToken' }]
    get:
      tags: [Invites]
      summary: Preview an invite
      security: []
      description: Public; invalid tokens return a 200 response with valid=false.
      responses:
        '200': { description: Invite preview, content: { application/json: { schema: { $ref: '#/components/schemas/InvitePreview' } } } }
  /invites/{token}/accept:
    parameters: [{ $ref: '#/components/parameters/InviteToken' }]
    post:
      tags: [Invites]
      summary: Accept an invite
      security: [{ sessionCookie: [] }]
      requestBody: { required: false, content: { application/json: { schema: { $ref: '#/components/schemas/AcceptInviteInput' } } } }
      responses:
        '200': { description: Joined member, content: { application/json: { schema: { $ref: '#/components/schemas/GroupMember' } } } }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '404': { $ref: '#/components/responses/NotFound' }
  /groups/{groupId}/invitations:
    parameters: [{ $ref: '#/components/parameters/GroupId' }]
    post:
      tags: [Invites]
      summary: Invite an existing user by email
      security: [{ sessionCookie: [] }]
      description: Requires group-admin access; sends an in-app notification.
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/InviteUserInput' } } } }
      responses:
        '204': { description: Invitation created }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
...
  /notifications:
    get:
      tags: [Notifications]
      summary: List notifications
      security: [{ sessionCookie: [] }]
      responses:
        '200': { description: Notifications, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/AppNotification' } } } } }
        '401': { $ref: '#/components/responses/Unauthenticated' }
...
  /notifications/{notificationId}/read:
    parameters: [{ $ref: '#/components/parameters/NotificationId' }]
    post:
      tags: [Notifications]
      summary: Mark one notification read
      security: [{ sessionCookie: [] }]
      responses:
        '204': { description: Notification marked read }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '404': { $ref: '#/components/responses/NotFound' }
  /notifications/read-all:
    post:
      tags: [Notifications]
      summary: Mark all notifications read
      security: [{ sessionCookie: [] }]
      responses:
        '204': { description: Notifications marked read }
        '401': { $ref: '#/components/responses/Unauthenticated' }
...
components:
  securitySchemes:
    sessionCookie:
      type: http
      scheme: bearer
      bearerFormat: opaque token
      description: Bearer token returned by /auth/register and /auth/login.
  parameters:
    GroupId: { name: groupId, in: path, required: true, schema: { $ref: '#/components/schemas/Id' } }
    MemberId: { name: memberId, in: path, required: true, schema: { $ref: '#/components/schemas/Id' } }
...
    NotificationId: { name: notificationId, in: path, required: true, schema: { $ref: '#/components/schemas/Id' } }
    InviteToken: { name: token, in: path, required: true, schema: { type: string } }
  responses:
    Unauthenticated: { description: Authentication required, content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } } }
    Forbidden: { description: Insufficient permissions, content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } } }
    NotFound: { description: Resource not found, content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } } }
    Conflict: { description: Conflict, content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } } }
    ValidationError: { description: Validation failed, content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } } }
  schemas:
    Id: { type: string }
    IsoDate: { type: string, format: date-time }
    CurrencyCode: { type: string, enum: [NGN] }
    Kobo: { type: integer, description: Integer amount in kobo (NGN minor units). }
    SplitMethod: { type: string, enum: [equal, custom, percentage, shares] }
    MemberRole: { type: string, enum: [admin, member] }
...
    Group:
      type: object
      required: [id, kind, name, description, ownerId, currency, inviteToken, createdAt, updatedAt]
      properties: { id: { $ref: '#/components/schemas/Id' }, kind: { type: string, enum: [group, quick] }, name: { type: string }, description: { type: string }, ownerId: { oneOf: [{ $ref: '#/components/schemas/Id' }, { type: 'null' }] }, currency: { $ref: '#/components/schemas/CurrencyCode' }, inviteToken: { oneOf: [{ type: string }, { type: 'null' }] }, createdAt: { $ref: '#/components/schemas/IsoDate' }, updatedAt: { $ref: '#/components/schemas/IsoDate' } }
    GroupMember:
      type: object
      required: [id, groupId, userId, displayName, role, joinedAt]
      properties: { id: { $ref: '#/components/schemas/Id' }, groupId: { $ref: '#/components/schemas/Id' }, userId: { oneOf: [{ $ref: '#/components/schemas/Id' }, { type: 'null' }] }, displayName: { type: string }, role: { $ref: '#/components/schemas/MemberRole' }, joinedAt: { $ref: '#/components/schemas/IsoDate' } }
    MemberDebtSummary:
      type: object
      required: [owes, owed, net]
      properties: { owes: { $ref: '#/components/schemas/Kobo' }, owed: { $ref: '#/components/schemas/Kobo' }, net: { $ref: '#/components/schemas/Kobo' } }
    GroupSummary:
      type: object
      required: [group, memberCount, expenseCount, totalSpent, viewerDebts, lastActivityAt]
      properties: { group: { $ref: '#/components/schemas/Group' }, memberCount: { type: integer }, expenseCount: { type: integer }, totalSpent: { $ref: '#/components/schemas/Kobo' }, viewerDebts: { oneOf: [{ $ref: '#/components/schemas/MemberDebtSummary' }, { type: 'null' }] }, lastActivityAt: { $ref: '#/components/schemas/IsoDate' } }
    CreateGroupInput:
      type: object
      required: [name]
      properties: { name: { type: string }, description: { type: string }, currency: { $ref: '#/components/schemas/CurrencyCode' }, memberNames: { type: array, items: { type: string } } }
    UpdateGroupInput:
      type: object
      minProperties: 1
      properties: { name: { type: string }, description: { type: string } }
    CreateQuickSplitInput:
      type: object
      required: [memberNames]
      properties: { name: { type: string }, memberNames: { type: array, minItems: 2, items: { type: string } } }
    AddMemberInput: { type: object, required: [displayName], properties: { displayName: { type: string } } }
    InviteToken: { type: object, required: [token], properties: { token: { type: string } } }
    InvitePreview:
      type: object
      required: [groupId, groupName, memberCount, valid]
      properties: { groupId: { $ref: '#/components/schemas/Id' }, groupName: { type: string }, memberCount: { type: integer }, valid: { type: boolean } }
    AcceptInviteInput: { type: object, properties: { displayName: { type: string } } }
    InviteUserInput: { type: object, required: [email], properties: { email: { type: string, format: email } } }
    ExpensePayerInput: { type: object, required: [memberId, amount], properties: { memberId: { $ref: '#/components/schemas/Id' }, amount: { $ref: '#/components/schemas/Kobo' } } }
...
    Receipt: { type: object, required: [id, expenseId, fileName, fileUrl, createdAt], properties: { id: { $ref: '#/components/schemas/Id' }, expenseId: { $ref: '#/components/schemas/Id' }, fileName: { type: string }, fileUrl: { type: string }, createdAt: { $ref: '#/components/schemas/IsoDate' } } }
    ExpensePayer: { $ref: '#/components/schemas/ExpensePayerInput' }
    ExpenseParticipant: { type: object, required: [memberId, allocationValue, calculatedAmount], properties: { memberId: { $ref: '#/components/schemas/Id' }, allocationValue: { type: number }, calculatedAmount: { $ref: '#/components/schemas/Kobo' } } }
    Expense:
      type: object
      required: [id, groupId, createdBy, description, totalAmount, currency, expenseDate, splitMethod, payers, participants, receipt, createdAt, updatedAt, deletedAt]
      properties: { id: { $ref: '#/components/schemas/Id' }, groupId: { $ref: '#/components/schemas/Id' }, createdBy: { $ref: '#/components/schemas/Id' }, description: { type: string }, totalAmount: { $ref: '#/components/schemas/Kobo' }, currency: { $ref: '#/components/schemas/CurrencyCode' }, expenseDate: { type: string, format: date }, splitMethod: { $ref: '#/components/schemas/SplitMethod' }, payers: { type: array, items: { $ref: '#/components/schemas/ExpensePayer' } }, participants: { type: array, items: { $ref: '#/components/schemas/ExpenseParticipant' } }, receipt: { oneOf: [{ $ref: '#/components/schemas/Receipt' }, { type: 'null' }] }, createdAt: { $ref: '#/components/schemas/IsoDate' }, updatedAt: { $ref: '#/components/schemas/IsoDate' }, deletedAt: { oneOf: [{ $ref: '#/components/schemas/IsoDate' }, { type: 'null' }] } }
...
    Settlement: { type: object, required: [id, groupId, fromMemberId, toMemberId, amount, status, paidAt, markedPaidBy, createdAt], properties: { id: { $ref: '#/components/schemas/Id' }, groupId: { $ref: '#/components/schemas/Id' }, fromMemberId: { $ref: '#/components/schemas/Id' }, toMemberId: { $ref: '#/components/schemas/Id' }, amount: { $ref: '#/components/schemas/Kobo' }, status: { type: string, enum: [paid] }, paidAt: { $ref: '#/components/schemas/IsoDate' }, markedPaidBy: { $ref: '#/components/schemas/Id' }, createdAt: { $ref: '#/components/schemas/IsoDate' } } }
    ViewerContext: { type: object, required: [memberId, isAdmin, canEditGroup, canManageMembers, canManageInvite], properties: { memberId: { oneOf: [{ $ref: '#/components/schemas/Id' }, { type: 'null' }] }, isAdmin: { type: boolean }, canEditGroup: { type: boolean }, canManageMembers: { type: boolean }, canManageInvite: { type: boolean } } }
    DebtEdge: { type: object, required: [fromMemberId, toMemberId, amount, expenseId, expenseDescription, expenseDate], properties: { fromMemberId: { $ref: '#/components/schemas/Id' }, toMemberId: { $ref: '#/components/schemas/Id' }, amount: { $ref: '#/components/schemas/Kobo' }, expenseId: { $ref: '#/components/schemas/Id' }, expenseDescription: { type: string }, expenseDate: { type: string, format: date } } }
    PairDebt: { type: object, required: [fromMemberId, toMemberId, grossAmount, settledAmount, outstandingAmount, edges], properties: { fromMemberId: { $ref: '#/components/schemas/Id' }, toMemberId: { $ref: '#/components/schemas/Id' }, grossAmount: { $ref: '#/components/schemas/Kobo' }, settledAmount: { $ref: '#/components/schemas/Kobo' }, outstandingAmount: { $ref: '#/components/schemas/Kobo' }, edges: { type: array, items: { $ref: '#/components/schemas/DebtEdge' } } } }
    MemberBalance: { type: object, required: [memberId, displayName, totalPaid, totalShare, netBalance], properties: { memberId: { $ref: '#/components/schemas/Id' }, displayName: { type: string }, totalPaid: { $ref: '#/components/schemas/Kobo' }, totalShare: { $ref: '#/components/schemas/Kobo' }, netBalance: { $ref: '#/components/schemas/Kobo' } } }
    GroupBalances: { type: object, required: [groupId, currency, members, debts], properties: { groupId: { $ref: '#/components/schemas/Id' }, currency: { $ref: '#/components/schemas/CurrencyCode' }, members: { type: array, items: { $ref: '#/components/schemas/MemberBalance' } }, debts: { type: array, items: { $ref: '#/components/schemas/PairDebt' } } } }
    GroupDetail: { type: object, required: [group, members, expenses, balances, settlements, activities, viewer], properties: { group: { $ref: '#/components/schemas/Group' }, members: { type: array, items: { $ref: '#/components/schemas/GroupMember' } }, expenses: { type: array, items: { $ref: '#/components/schemas/Expense' } }, balances: { $ref: '#/components/schemas/GroupBalances' }, settlements: { type: array, items: { $ref: '#/components/schemas/Settlement' } }, activities: { type: array, items: { $ref: '#/components/schemas/Activity' } }, viewer: { $ref: '#/components/schemas/ViewerContext' } } }
    Activity: { type: object, required: [id, groupId, actorId, actorName, action, entityType, entityId, summary, createdAt], properties: { id: { $ref: '#/components/schemas/Id' }, groupId: { $ref: '#/components/schemas/Id' }, actorId: { oneOf: [{ $ref: '#/components/schemas/Id' }, { type: 'null' }] }, actorName: { type: string }, action: { type: string }, entityType: { type: string, enum: [expense, member, group, settlement, invite] }, entityId: { oneOf: [{ $ref: '#/components/schemas/Id' }, { type: 'null' }] }, summary: { type: string }, createdAt: { $ref: '#/components/schemas/IsoDate' } } }
    AppNotification: { type: object, required: [id, userId, groupId, type, title, message, readAt, createdAt], properties: { id: { $ref: '#/components/schemas/Id' }, userId: { $ref: '#/components/schemas/Id' }, groupId: { oneOf: [{ $ref: '#/components/schemas/Id' }, { type: 'null' }] }, type: { type: string, enum: [expense_created, expense_edited, expense_deleted, debt_marked_paid, group_invitation] }, title: { type: string }, message: { type: string }, readAt: { oneOf: [{ $ref: '#/components/schemas/IsoDate' }, { type: 'null' }] }, createdAt: { $ref: '#/components/schemas/IsoDate' } } }
...
    ApiError: { type: object, properties: { code: { type: string, enum: [validation, not_found, forbidden, unauthenticated, conflict] }, message: { type: string }, issues: { type: array, items: { $ref: '#/components/schemas/ValidationIssue' } } } }
    ValidationIssue: { type: object, required: [field, message], properties: { field: { type: string, enum: [description, totalAmount, expenseDate, payers, participants, allocation] }, message: { type: string } } }